Moved patching of binding to a  function called after injection
