+++ packages/realm/src/app-services/NetworkTransport.ts
@@ -16,7 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-import { binding, extendDebug, network, toFetchArgs } from "../internal";
+import { binding, extendDebug, network } from "../internal";
 import type { Headers } from "@realm/fetch";
 
 const debug = extendDebug("network");
@@ -32,7 +32,7 @@ function flattenHeaders(headers: Headers) {
 /** @internal */
 export function createNetworkTransport(fetch = network.fetch) {
   return binding.Helpers.makeNetworkTransport((request, callback) => {
-    const [url, init] = toFetchArgs(request);
+    const [url, init] = binding.toFetchArgs(request);
     debug("requesting %s %O", url, init);
     fetch(url, init).then(
       async (response) => {
